Simplify plot task inheritance.
