+++ columnflow/tasks/framework/decorators.py
@@ -1,0 +1,49 @@
+# coding: utf-8
+
+"""
+Custom law task method decorators.
+"""
+
+import law
+
+
+@law.decorator.factory(accept_generator=True)
+def view_output_plots(fn, opts, task, *args, **kwargs):
+    def before_call():
+        return None
+
+    def call(state):
+        return fn(task, *args, **kwargs)
+
+    def after_call(state):
+        view_cmd = getattr(task, "view_cmd", None)
+        if not view_cmd or view_cmd == law.NO_STR:
+            return
+
+        # prepare the view command
+        if "{}" not in view_cmd:
+            view_cmd += " {}"
+
+        # collect all paths to view
+        view_paths = []
+        outputs = law.util.flatten(task.output())
+        while outputs:
+            output = outputs.pop(0)
+            if isinstance(output, law.TargetCollection):
+                outputs.extend(output._flat_target_list)
+                continue
+            if not getattr(output, "path", None):
+                continue
+            if output.path.endswith((".pdf", ".png")):
+                if not isinstance(output, law.LocalTarget):
+                    task.logger.warning(f"cannot show non-local plot at '{output.path}'")
+                    continue
+                elif output.path not in view_paths:
+                    view_paths.append(output.path)
+
+        # loop through paths and view them
+        for path in view_paths:
+            task.publish_message("showing {}".format(path))
+            law.util.interruptable_popen(view_cmd.format(path), shell=True, executable="/bin/bash")
+
+    return before_call, call, after_call
