+++ columnflow/tasks/framework/plotting.py
@@ -107,51 +107,6 @@ class PlotBase(AnalysisTask):
         return kwargs
 
 
-@law.decorator.factory(accept_generator=True)
-def view_output_plots(fn, opts, task, *args, **kwargs):
-    def before_call():
-        return None
-
-    def call(state):
-        return fn(task, *args, **kwargs)
-
-    def after_call(state):
-        view_cmd = getattr(task, "view_cmd", None)
-        if not view_cmd or view_cmd == law.NO_STR:
-            return
-
-        # prepare the view command
-        if "{}" not in view_cmd:
-            view_cmd += " {}"
-
-        # collect all paths to view
-        view_paths = []
-        outputs = law.util.flatten(task.output())
-        while outputs:
-            output = outputs.pop(0)
-            if isinstance(output, law.TargetCollection):
-                outputs.extend(output._flat_target_list)
-                continue
-            if not getattr(output, "path", None):
-                continue
-            if output.path.endswith((".pdf", ".png")):
-                if not isinstance(output, law.LocalTarget):
-                    task.logger.warning(f"cannot show non-local plot at '{output.path}'")
-                    continue
-                elif output.path not in view_paths:
-                    view_paths.append(output.path)
-
-        # loop through paths and view them
-        for path in view_paths:
-            task.publish_message("showing {}".format(path))
-            law.util.interruptable_popen(view_cmd.format(path), shell=True, executable="/bin/bash")
-
-    return before_call, call, after_call
-
-
-PlotBase.view_output_plots = view_output_plots
-
-
 class PlotBase1D(PlotBase):
     """
     Base class for plotting tasks creating 1-dimensional plots.
@@ -284,12 +239,6 @@ class ProcessPlotSettingMixin(
                 params["process_settings"] = groups[params["process_settings"][0][0]]
 
         return params
-
-    def store_parts(self):
-        parts = super().store_parts()
-        part = f"datasets_{self.datasets_repr}"
-        parts.insert_before("version", "plot", part)
-        return parts
 
 
 class VariablePlotSettingMixin(
