+++ columnflow/tasks/plotting.py
@@ -5,6 +5,7 @@ Tasks to plot different types of histograms.
 """
 
 from collections import OrderedDict
+from abc import abstractmethod
 
 import law
 import luigi
@@ -17,14 +18,15 @@ from columnflow.tasks.framework.mixins import (
 from columnflow.tasks.framework.plotting import (
     PlotBase, PlotBase1D, PlotBase2D, ProcessPlotSettingMixin, VariablePlotSettingMixin,
 )
+from columnflow.tasks.framework.decorators import view_output_plots
 from columnflow.tasks.framework.remote import RemoteWorkflow
 from columnflow.tasks.histograms import MergeHistograms, MergeShiftedHistograms
 from columnflow.util import DotDict, dev_sandbox, dict_add_strict
 
 
 class PlotVariablesBase(
-    ShiftTask,
     VariablePlotSettingMixin,
+    ProcessPlotSettingMixin,
     MLModelsMixin,
     ProducersMixin,
     SelectorStepsMixin,
@@ -35,21 +37,14 @@ class PlotVariablesBase(
     law.LocalWorkflow,
     RemoteWorkflow,
 ):
-    exclude_index = True
-
     sandbox = dev_sandbox("bash::$CF_BASE/sandboxes/venv_columnar.sh")
 
-    shifts = set(MergeHistograms.shifts)
+    exclude_index = True
 
-    # default upstream dependency task classes
-    dep_MergeHistograms = MergeHistograms
-
-    def create_branch_map(self):
-        return [
-            DotDict({"category": cat_name, "variable": var_name})
-            for var_name in self.variables
-            for cat_name in self.categories
-        ]
+    def store_parts(self):
+        parts = super().store_parts()
+        parts.insert_before("version", "plot", f"datasets_{self.datasets_repr}")
+        return parts
 
     def workflow_requires(self, only_super: bool = False):
         reqs = super().workflow_requires()
@@ -60,29 +55,17 @@ class PlotVariablesBase(
 
         return reqs
 
-    def requires(self):
-        return {
-            d: self.dep_MergeHistograms.req(
-                self,
-                dataset=d,
-                branch=-1,
-                _exclude={"branches"},
-                _prefer_cli={"variables"},
-            )
-            for d in self.datasets
-        }
-
-    def output(self):
-        b = self.branch_data
-        return [
-            self.target(name)
-            for name in self.get_plot_names(f"plot__proc_{self.processes_repr}__cat_{b.category}__var_{b.variable}")
-        ]
+    @abstractmethod
+    def get_plot_shifts(self):
+        return
 
     @law.decorator.log
-    @PlotBase.view_output_plots
+    @view_output_plots
     def run(self):
         import hist
+
+        # get the shifts to extract and plot
+        plot_shifts = law.util.make_list(self.get_plot_shifts())
 
         # prepare config objects
         variable_tuple = self.variable_tuples[self.branch_data.variable]
@@ -106,11 +89,6 @@ class PlotVariablesBase(
                 dataset_inst = self.config_inst.get_dataset(dataset)
                 h_in = inp["collection"][0].targets[self.branch_data.variable].load(formatter="pickle")
 
-                # sanity checks
-                n_shifts = len(h_in.axes["shift"])
-                if n_shifts != 1:
-                    raise Exception(f"shift axis is supposed to only contain 1 bin, found {n_shifts}")
-
                 # loop and extract one histogram per process
                 for process_inst in process_insts:
                     # skip when the dataset is already known to not contain any sub process
@@ -132,10 +110,15 @@ class PlotVariablesBase(
                             for c in leaf_category_insts
                             if c.id in h.axes["category"]
                         ],
+                        "shift": [
+                            hist.loc(s.id)
+                            for s in plot_shifts
+                            if s.id in h.axes["shift"]
+                        ],
                     }]
 
                     # axis reductions
-                    h = h[{"process": sum, "category": sum, "shift": sum}]
+                    h = h[{"process": sum, "category": sum}]
 
                     # add the histogram
                     if process_inst in hists:
@@ -178,21 +161,58 @@ class PlotVariablesBase(
                 outp.dump(fig, formatter="mpl")
 
 
-class PlotVariables1D(
+class PlotVariablesBaseSingleShift(
+    ShiftTask,
     PlotVariablesBase,
-    PlotBase1D,
-    ProcessPlotSettingMixin,
 ):
+    shifts = set(MergeHistograms.shifts)
 
+    # default upstream dependency task classes
+    dep_MergeHistograms = MergeHistograms
+
+    exclude_index = True
+
+    def create_branch_map(self):
+        return [
+            DotDict({"category": cat_name, "variable": var_name})
+            for var_name in self.variables
+            for cat_name in self.categories
+        ]
+
+    def requires(self):
+        return {
+            d: self.dep_MergeHistograms.req(
+                self,
+                dataset=d,
+                branch=-1,
+                _exclude={"branches"},
+                _prefer_cli={"variables"},
+            )
+            for d in self.datasets
+        }
+
+    def output(self):
+        b = self.branch_data
+        return [
+            self.target(name)
+            for name in self.get_plot_names(f"plot__proc_{self.processes_repr}__cat_{b.category}__var_{b.variable}")
+        ]
+
+    def get_plot_shifts(self):
+        return [self.shift_inst]
+
+
+class PlotVariables1D(
+    PlotVariablesBaseSingleShift,
+    PlotBase1D,
+):
     pass
 
 
 class PlotVariables2D(
-    PlotVariablesBase,
+    PlotVariablesBaseSingleShift,
     PlotBase2D,
-    ProcessPlotSettingMixin,
 ):
-
     pass
 
 
@@ -200,7 +220,6 @@ class PlotVariablesPerProcess2D(
     law.WrapperTask,
     PlotVariables2D,
 ):
-
     # force this one to be a local workflow
     workflow = "local"
 
@@ -211,48 +230,26 @@ class PlotVariablesPerProcess2D(
         }
 
 
-class PlotShiftedVariables1D(
-    VariablePlotSettingMixin,
+class PlotVariablesBaseMultiShifts(
     ShiftSourcesMixin,
-    MLModelsMixin,
-    ProducersMixin,
-    SelectorStepsMixin,
-    CalibratorsMixin,
-    CategoriesMixin,
-    PlotBase1D,
-    ProcessPlotSettingMixin,
-    law.LocalWorkflow,
-    RemoteWorkflow,
+    PlotVariablesBase,
 ):
-
     legend_title = luigi.Parameter(
         default=law.NO_STR,
         significant=False,
         description="sets the title of the legend; when empty and only one process is present in "
         "the plot, the process_inst label is used; empty default",
     )
-    # overriding the default plot function
     plot_function_1d = luigi.Parameter(
         default="columnflow.plotting.example.plot_shifted_variable",
         significant=False,
         description="name of the 1d plot function; default: 'columnflow.plotting.example.plot_shifted_variable'",
     )
 
-    sandbox = dev_sandbox("bash::$CF_BASE/sandboxes/venv_columnar.sh")
-
     # default upstream dependency task classes
     dep_MergeShiftedHistograms = MergeShiftedHistograms
 
-    def get_plot_parameters(self):
-        # convert parameters to usable values during plotting
-        params = super().get_plot_parameters()
-        dict_add_strict(params, "legend_title", None if self.legend_title == law.NO_STR else self.legend_title)
-        return params
-
-    def store_parts(self):
-        parts = super().store_parts()
-        parts["plot"] += f"__shifts_{self.shift_sources_repr}"
-        return parts
+    exclude_index = True
 
     def create_branch_map(self):
         return [
@@ -261,14 +258,6 @@ class PlotShiftedVariables1D(
             for cat_name in self.categories
             for source in self.shift_sources
         ]
-
-    def workflow_requires(self, only_super: bool = False):
-        reqs = super().workflow_requires()
-        if only_super:
-            return reqs
-
-        reqs["merged_hists"] = self.requires_from_branch()
-        return reqs
 
     def requires(self):
         return {
@@ -284,99 +273,37 @@ class PlotShiftedVariables1D(
 
     def output(self):
         b = self.branch_data
-        return [self.target(name) for name in self.get_plot_names
-        (
-            f"plot__proc_{self.processes_repr}__unc_{b.shift_source}__cat_{b.category}__var_{b.variable}"
-        )]
+        return [
+            self.target(name)
+            for name in self.get_plot_names(
+                f"plot__proc_{self.processes_repr}__unc_{b.shift_source}__cat_{b.category}__var_{b.variable}",
+            )
+        ]
 
-    @law.decorator.log
-    @PlotBase.view_output_plots
-    def run(self):
-        import hist
-
-        # prepare config objects
-        variable_inst = self.config_inst.get_variable(self.branch_data.variable)
-        category_inst = self.config_inst.get_category(self.branch_data.category)
-        shift_insts = [
+    def get_plot_shifts(self):
+        return [
             self.config_inst.get_shift(s) for s in
             ["nominal", f"{self.branch_data.shift_source}_up", f"{self.branch_data.shift_source}_down"]
         ]
-        leaf_category_insts = category_inst.get_leaf_categories() or [category_inst]
-        process_insts = list(map(self.config_inst.get_process, self.processes))
-        sub_process_insts = {
-            proc: [sub for sub, _, _ in proc.walk_processes(include_self=True)]
-            for proc in process_insts
-        }
 
-        # histogram data per process
-        hists = OrderedDict()
-        outputs = self.output()
+    def get_plot_parameters(self):
+        # convert parameters to usable values during plotting
+        params = super().get_plot_parameters()
+        dict_add_strict(params, "legend_title", None if self.legend_title == law.NO_STR else self.legend_title)
+        return params
 
-        with self.publish_step(f"plotting {variable_inst.name} in {category_inst.name}"):
-            for dataset, inp in self.input().items():
-                dataset_inst = self.config_inst.get_dataset(dataset)
-                h_in = inp["collection"][0].targets[variable_inst.name].load(formatter="pickle")
 
-                # loop and extract one histogram per process
-                for process_inst in process_insts:
-                    # skip when the dataset is already known to not contain any sub process
-                    if not any(map(dataset_inst.has_process, sub_process_insts[process_inst])):
-                        continue
-
-                    # work on a copy
-                    h = h_in.copy()
-
-                    # axis selections
-                    h = h[{
-                        "process": [
-                            hist.loc(p.id)
-                            for p in sub_process_insts[process_inst]
-                            if p.id in h.axes["process"]
-                        ],
-                        "category": [
-                            hist.loc(c.id)
-                            for c in leaf_category_insts
-                            if c.id in h.axes["category"]
-                        ],
-                        "shift": [
-                            hist.loc(s.id)
-                            for s in shift_insts
-                            if s.id in h.axes["shift"]
-                        ],
-                    }]
-
-                    # axis reductions
-                    h = h[{"process": sum, "category": sum}]
-
-                    # add the histogram
-                    if process_inst in hists:
-                        hists[process_inst] += h
-                    else:
-                        hists[process_inst] = h
-
-            # there should be hists to plot
-            if not hists:
-                raise Exception("no histograms found to plot")
-
-            # call the plot function once
-            fig = self.call_plot_func(
-                self.plot_function,
-                hists=hists,
-                config_inst=self.config_inst,
-                variable_inst=variable_inst,
-                **self.get_plot_parameters(),
-            )
-
-            # save the plot
-            for outp in outputs:
-                outp.dump(fig, formatter="mpl")
+class PlotShiftedVariables1D(
+    PlotVariablesBaseMultiShifts,
+    PlotBase1D,
+):
+    pass
 
 
 class PlotShiftedVariablesPerProcess1D(
     law.WrapperTask,
     PlotShiftedVariables1D,
 ):
-
     # force this one to be a local workflow
     workflow = "local"
